VFS: Separate implementation from header
This has lived for far too long in a single file.
